# MultiCFSnapshot Refactor - separate multiget key range info from CFD & superversion info